Warning code refactor
